Implement persistent theme toggle
