+++ package.json
@@ -18,9 +18,13 @@
     "@types/react-dom": "18.2.3",
     "eslint": "8.39.0",
     "eslint-config-next": "13.4.0",
+    "js-cookie": "^3.0.5",
     "next": "13.4.0",
     "react": "18.2.0",
     "react-dom": "18.2.0",
     "typescript": "5.0.4"
+  },
+  "devDependencies": {
+    "@types/js-cookie": "^3.0.3"
   }
 }
