+++ src/app/components/ThemeToggle/index.ts
@@ -1,0 +1,3 @@
+import ThemeToggle from "./toggle";
+
+export default ThemeToggle;
